Fix metadata export by moving home page to component
